write cache files when enabling formats and read them at boot
A cache directory now contains the digested output from act_enable()
ready to be written to $PROCDIR/register.

act_enable writes cache entries even in the cases it does not actually
write itself to $PROCDIR/register (already loaded or no PROCDIR).

At boot, the cache is loaded and compared against the database. If it
is not successful or complete, the full update-binfmts --enable is run
as a fallback.